Fix strlen redundancy
Use of strlen requires full traversal of the string on every call. In calculator.c, strlen was used multiple times on the same string, which, when dealing with long tokens, could result in a performance bottleneck. This commit fixes the issue by storing necessary string lengths in a variable, which is then reused when needed.